Set or Retrieve Default Plot Parameters for Model Diagnostic Plots

Initializes or updates a list of plot element parameters for use in model diagnostic plotting functions. If any required parameter is missing or invalid, a default value will be supplied.

Usage

lm_plot.parms(parms = list())

Arguments

parms

A list of plot parameters. Any missing or invalid entries are replaced with defaults.

Value

A list of plot element parameters with defaults filled in for any missing or invalid entries.

Details

The returned list contains parameters for points (size, color, shape), lines (type, color, size), options for plot features, and Cook's distance/Influence contours. These are used by other lm_plot.* functions to control plot appearance and annotation. Key defaults include:

pts: Properties for points (size, color, shape, outlier flags)

lins: Properties for lines (type, color, size)

opt: Logical options for plot features (e.g., time series, prediction interval, test statistics)

cook: Cook's distance contour settings (points, levels, line type)

infl: Influence line settings (line type)

See function code for full list of available settings and default values.

See also

Examples

# Retrieve default parameters
parms <- lm_plot.parms()
# Set custom color for regular points
parms <- lm_plot.parms(list(pts = list(colr = list(reg = "blue"))))